requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
idna==2.10
ImageHash==4.2.0
multidict==5.1.0
numpy==1.19.5
Pillow==8.1.0
PyWavelets==1.1.1
requests==2.25.1
scipy==1.6.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
